Added a prototype minimap for QGraphicsScene (#24)
* Added a prototype minimap for QGraphicsScene

* Add InfiniteScrollGraphicsViewBase

* Changed the boundary values of sceneRect for InfiniteScrollGraphicsViewBase.

Changed the range of values to (INT_MIN, INT_MAX), because scrolling and centerOn() started working incorrectly with a larger range (I suspect that they are not the only ones. I think this behavior occurs because of the scrollbar limits in QGraphicsView)

* Redesigned the architecture of the SceneMinimap

Removed the division into wrapper and -impl versions of the class. Now the minimap is a separate widget and no longer inherits the QGraphicsView functionality in the implementation, but simply renders the QGraphicsView passed to it, as it should.

* Implement a valid implementation of SceneMinimap::mapFromScene()

* Rendering the target view rect in m_targetView->viewport()->geometry() instead of m_targetView->geometry()

* Implement SceneMinimap::mouseMoveEvent()

* Moved the widgets to a separate module

* Add SceneViewBase

Currently this class allows track changes in the transform matrix of the current view.

* Add a minimap update when redrawing the target view

* Separate minimap painting steps into own private methods

* Add a customization point for the target view rectangle color

* Add a fill color alignment calculation based on the color of the scene background color

* Separate the common parts of SceneMinimap::mapFromScene() and SceneMinimap::mapToScene() into separate private methods

* Add an action for creating an external minimap for WhiteboardScene to the debugging menu

* Small improvements

Fix the position of the minimap in WhiteboardView, display the minimap only when the scene does not fit in the view rect