Redesign the rendering of ToolItemBase. (#27)
* Redesigned the rendering of ToolItemBase.

Now, by default, ToolItemBase will draw and calculate the limit rect using its shape().  Also, the ItemIsSelectable flag is set by default.
As a result, all derived ToolItemBase classes now define shape() instead of the custom getLimitRect() and onPaint()

* Split a monolith CMakeLists.txt into modules

* Disable CXX_SCAN_FOR_MODULES for 'boardroom' target

* Add 'm_thickness' param into account to DrawingItemBase::getLimitRect()

* Fixed the shape() methods for EllispeItem and RectangleItem

The problem was that before creating shape(), it is necessary to call moveTo() to the starting position, even if we create the shape according to all known coordinates.

* Restore the thickness of DrawingItemBase